chore(frontend): fix deps update
